requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.21.0
PyYAML>=6.0.0
python-dotenv>=1.0.0
openai>=1.0.0

# PDF processing
pdfplumber>=0.7.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0

# CLI and UI
typer>=0.9.0
rich>=12.0.0
click>=8.0.0

# Google Sheets API
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
gspread>=5.0.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950 